backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
openai>=1.50.0
pyyaml>=6.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
chromadb>=1.5.0
